requirements.txt
fastapi==0.103.0
uvicorn[standard]==0.23.2
streamlit==1.26.0
tensorflow==2.16.1
torch==2.1.0
ultralytics==8.1.114
transformers==4.35.0
numpy==1.26.1
Pillow==10.1.0
requests==2.31.0
scikit-image==0.21.0
azure-storage-blob==12.25.0
scikit-learn==1.3.0
